+++ src/components/feed/GroupsActivitiesViews.tsx
@@ -4,7 +4,7 @@ import {
   ACTIVITY_FEED_VIEWS,
   ActivityFeedViewType,
 } from "@/src/utils/constants";
-import useGetAllSearchParams from "@/src/utils/helpers/useGetAllSearchParams";
+import useGetAllSearchParams from "@/src/utils/helpers/getAllSearchParams";
 import Image from "next/image";
 import Link from "next/link";
 import React, { useState } from "react";
